Token Renew deleted Cleaned a little bit
