+++ Downloader/ChunkManager.cs
@@ -1,5 +1,4 @@
-﻿using System.Text.RegularExpressions;
-using Uplay.Download;
+﻿using Uplay.Download;
 using File = Uplay.Download.File;
 
 namespace Downloader
@@ -8,15 +7,11 @@ namespace Downloader
     {
         public static List<File> RemoveNonEnglish(Manifest parsedManifest)
         {
-            List<Chunk> reqChunks = parsedManifest.Chunks.Where(x => String.IsNullOrEmpty(x.Language) && x.Type == Chunk.Types.ChunkType.Required).ToList(); // filter out non-English
-            List<Chunk> optChunks = parsedManifest.Chunks.Where(x => String.IsNullOrEmpty(x.Language) && x.Type == Chunk.Types.ChunkType.Optional).ToList(); // filter out non-English
+            List<Chunk> chunks = parsedManifest.Chunks.Where(x => String.IsNullOrEmpty(x.Language)).ToList();
 
             List<File> files = new();
 
-            foreach (var chunk in reqChunks)
-                files = files.Concat(chunk.Files).ToList();
-
-            foreach (var chunk in optChunks)
+            foreach (var chunk in chunks)
                 files = files.Concat(chunk.Files).ToList();
 
             return files;
@@ -35,26 +30,22 @@ namespace Downloader
 
         public static List<File> AddLanguage(Manifest parsedManifest, string Lang)
         {
-            List<Chunk> reqChunks = parsedManifest.Chunks.Where(x => x.Language.Equals(Lang) && x.Type == Chunk.Types.ChunkType.Required).ToList();
-            List<Chunk> optChunks = parsedManifest.Chunks.Where(x => x.Language.Equals(Lang) && x.Type == Chunk.Types.ChunkType.Optional).ToList();
+            List<Chunk> chunks = parsedManifest.Chunks.Where(x => x.Language.Equals(Lang)).ToList();
 
             List<File> files = new();
 
-            foreach (var chunk in reqChunks)
-                files = files.Concat(chunk.Files).ToList();
-
-            foreach (var chunk in optChunks)
+            foreach (var chunk in chunks)
                 files = files.Concat(chunk.Files).ToList();
 
             return files;
         }
 
-        public static List<File> RemoveSkipFiles(List<File> files, List<string> skip_files)
+        public static void RemoveSkipFiles(List<string> skip_files)
         {
             List<File> to_remove = new();
             if (skip_files.Count != 0)
             {
-                foreach (var file in files)
+                foreach (var file in DLWorker.Config.FilesToDownload)
                 {
                     foreach (var skip in skip_files)
                     {
@@ -66,26 +57,28 @@ namespace Downloader
                         if (skip.StartsWith("regex:"))
                         {
                             var rgx = new Regex(skip.Substring(6), RegexOptions.Compiled | RegexOptions.IgnoreCase);
+                            var m = rgx.Match(file.Name);
+
+                            if (!m.Success)
+                                to_remove.Add(file);
                         }
                         */
                     }
                 }
                 foreach (var remove in to_remove)
                 {
-                    files.Remove(remove);
+                    DLWorker.Config.FilesToDownload.Remove(remove);
                 }
 
             }
-
-            return files;
         }
 
-        public static List<File> AddDLOnlyFiles(List<File> files, List<string> add_files)
+        public static List<File> AddDLOnlyFiles(List<string> add_files)
         {
             List<File> output = new();
             if (add_files.Count != 0)
             {
-                foreach (var file in files)
+                foreach (var file in DLWorker.Config.FilesToDownload)
                 {
                     foreach (var add in add_files)
                     {
